Added --filter flag
Added a new option --filter (-f) that allows the program to be used as a
unix filter. This means it takes input from stdin and prints the output to
stdout. This is helpful to use it as a git filter, so that cbproj files
can be automatically processed before being committed to git.